+++ extensions/gii/CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 gii extension Change Log
 2.0.0-rc under development
 --------------------------
 
-- no changes in this release.
+- Enh #3088: The gii module will manage their own URL rules now (qiangxue)
 
 
 2.0.0-beta April 13, 2014